requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Machine Learning & AI
torch==2.0.1
torchvision==0.15.2
sentence-transformers==2.2.2
scikit-learn==1.3.0

# Computer Vision & OCR
opencv-python==4.8.0.76
pytesseract==0.3.10
easyocr==1.7.0
Pillow==10.0.0

# Vector Database
pinecone

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Utilities
python-dotenv==1.0.0